backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
pillow
cloudinary
django-cloudinary-storage
stripe
paypalrestsdk
